move modeling functions to utils for generic use
