+++ .gitignore
@@ -1,5 +1,5 @@
 .DS_Store
+.env
 code/venv/*
 code/__pycache__/*
-code/models/__pycache__/*
 code/utils/__pycache__/*
